Introduce FixtureUtils for centralized FixtureMonkey configuration in tests
